Remove last traces of `divert()`
As with previous commits, testing this is a bit "meh", but consists of:

- `yarn test`
- Using the local copy of the generator to create a new theme called
  "test"; ie. `yo ./packages/generator-liferay-theme`
- `cd test-theme` and change this line in gulpfile.js to make it use
  the local tasks package:

    - var liferayThemeTasks = require('liferay-theme-tasks');
    + var liferayThemeTasks = require('../packages/liferay-theme-tasks');

- Run the "build" task; ie. `node_modules/.bin/gulp build`
