+++ packages/generator-liferay-theme/generators/app/__tests__/index.test.js
@@ -7,7 +7,6 @@ const path = require('path');
 const sinon = require('sinon');
 const helpers = require('yeoman-generator').test;
 
-const divert = require('liferay-theme-tasks/lib/divert');
 const liferayThemeApp = require('../index');
 
 chai.use(require('chai-fs'));
