+++ packages/liferay-theme-tasks/tasks/build.js
@@ -9,9 +9,9 @@ const replace = require('gulp-replace-task');
 const through = require('through2');
 const PluginError = require('plugin-error');
 
-const divert = require('../lib/divert');
 const lfrThemeConfig = require('../lib/liferay_theme_config');
 const lookAndFeelUtil = require('../lib/look_and_feel_util');
+const lookup = require('../lib/lookup');
 const normalize = require('../lib/normalize');
 const themeUtil = require('../lib/util');
 
@@ -86,9 +86,7 @@ module.exports = function(options) {
 	});
 
 	gulp.task('build:base', function() {
-		const sourceFiles = divert('dependencies').getBaseThemeDependencies(
-			process.cwd()
-		);
+		const sourceFiles = lookup('baseThemeDependencies')();
 
 		return gulp
 			.src(sourceFiles)
